+++ Algoritm/Algoritm.csproj
@@ -11,8 +11,4 @@
     <PackageReference Include="BenchmarkDotNet" Version="0.13.1" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="BenchmarkRunner\ResaltPointEXP\" />
-  </ItemGroup>
-
 </Project>
