+++ Algoritm/ExitClass.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Algoritm
+{
+    internal class ExitClass : Ilesson
+    {
+        public int id => 9;
+
+        public string Descprition => "Exit";
+
+        public void RUN()
+        {
+            Environment.Exit(0);
+        }
+    }
+}
