+++ Algoritm/Lesson2/LInkL/ILinkedList.cs
