+++ Algoritm/Lesson2/LInkL/LinkedList.cs
