+++ Algoritm/Lesson2/LInkL/LinkedListNode.cs
