+++ Algoritm/Lesson3/BenchmarkRunner/PointClassDouble.cs
