+++ Algoritm/Lesson3/BenchmarkRunner/PointStructDouble.cs
